re-generate uploader.rb
