+++ app/models/captured_image.rb
@@ -1,4 +1,6 @@
 class CapturedImage < ActiveRecord::Base
+  mount_uploader :content, ImageUploader
+
   enum status: %i(main sub)
 
   belongs_to :prototype
